generateTemplateFilesCommandLine and output.overwrite
