+++ examples/package.json
@@ -6,7 +6,7 @@
     "---------- Normal Example -------------------------------------------": "",
     "generate": "node ./tools/generate.js",
     "---------- Command Line Example -------------------------------------": "",
-    "commandline": "node ./tools/generate.js angular-ngrx-store __name__=some-name __model__=some-other-name --outputpath=./src/here --overwrite"
+    "commandline": "node ./tools/commandLine.js angular-ngrx-store __name__=some-name __model__=some-other-name --outputpath=./src/here --overwrite"
   },
   "devDependencies": {
     "file-name": "0.1.0",
