+++ src/index.ts
@@ -19,9 +19,12 @@ export type IReplacerSlotQuestion = IReplacerSlotQuestionDefault;
  * Main method to create your template files. Accepts an array of `IConfigItem` items.
  */
 export function generateTemplateFiles(data: IConfigItem[]): Promise<void> {
-    if (GenerateTemplateFiles.isCommandLine) {
-        return new GenerateTemplateFiles().commandLine(data);
-    }
-
     return new GenerateTemplateFiles().generate(data);
+}
+
+/**
+ * Main method to create your template files with command line.
+ */
+export function generateTemplateFilesCommandLine(data: IConfigItem[]): Promise<void> {
+    return new GenerateTemplateFiles().commandLine(data);
 }
