+++ src/utilities/GenerateTemplateFiles.commandline.test.ts
@@ -1,0 +1,3 @@
+describe('GenerateTemplateFiles - Command Line', () => {
+    //
+});
